IAA/Projeto_IAA_oauth/requirements.txt
Flask==1.1.2
Authlib==1.0.1
requests
Flask-SocketIO==5.1.2
cryptography==3.4.7
tabulate